Develop

+++ src/app/api/survey/[id]/route.ts
@@ -1,27 +1,26 @@
 import connectMongoDB from "@/lib/mongodb";
 import Survey from "@/models/surveys";
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import { ZodError } from "zod";
 import { headers } from "next/headers";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/lib/auth";
 
 export async function GET(
-  request: Request,
+  _request: Request,
   { params: { id } }: { params: { id: string } }
 ) {
   const session = await getServerSession(authOptions);
   const discordKey = headers().get("discord-key");
-  const user_id = session?.user?.id;
+
+  // Allow access to all logged in users
+  if (!(session || (discordKey && discordKey === process.env.BOT_KEY))) {
+    return new Response("401");
+  }
 
   try {
     await connectMongoDB();
     const survey = await Survey.findOne({ discordId: id });
-    if (
-      (!session && !discordKey && user_id !== survey.discordId) ||
-      (discordKey && discordKey !== process.env.BOT_KEY)
-    )
-      return new Response("401");
     return NextResponse.json({ survey }, { status: 200 });
   } catch (error) {
     if (error instanceof ZodError)
@@ -31,16 +30,26 @@ export async function GET(
   }
 }
 
-export async function DELETE(
-  request: Request,
-  { params: { id } }: { params: { id: string } }
-) {
-  const discordKey = headers().get("discord-key");
+export async function DELETE(request: Request) {
+  const discordKey = request.headers.get("discord-key");
 
-  // if (!discordKey || discordKey !== process.env.BOT_KEY)
-  //   return new Response("401");
+  // Allow access only to the Discord Bot
+  if (!discordKey || discordKey !== process.env.BOT_KEY)
+    return new Response("401");
 
-  await connectMongoDB();
-  await Survey.findByIdAndDelete(id);
-  return NextResponse.json({ message: "Survey deleted" }, { status: 200 });
+  const { searchParams } = new URL(request.url);
+  const id = searchParams.get("id");
+
+  if (!id) {
+    return NextResponse.json({ message: "ID is required" }, { status: 400 });
+  }
+
+  try {
+    await connectMongoDB();
+    await Survey.findByIdAndDelete(id);
+    return NextResponse.json({ message: "Survey deleted" }, { status: 200 });
+  } catch (error) {
+    if (error instanceof Error)
+      return NextResponse.json({ message: error.message }, { status: 500 });
+  }
 }
